feat: implement SimpleLogger
