+++ jobs/.eslintrc.json
@@ -42,6 +42,7 @@
     "camelcase": "off",
     "@typescript-eslint/explicit-module-boundary-types": "off",
     "no-shadow": "off",
+    "@typescript-eslint/no-empty-interface": "off",
     "@typescript-eslint/no-shadow": ["error"],
     "@typescript-eslint/no-unsafe-call": "off",
     "class-methods-use-this": "off"
