+++ jobs/src/types/event.ts
@@ -1,0 +1,14 @@
+import { ILogger } from '../util/logger';
+
+export interface EventContext {
+  logger: ILogger;
+}
+
+export interface EventOptions {
+  name: string;
+  buildString?: string;
+}
+
+export type EventHandler = (
+  context: EventContext
+) => Promise<[number, AnyJson]>;
