--- jobs/src/types/scheduledEvent.ts
@@ -1,14 +1,0 @@
-import Logger from '../util/logger';
-
-export interface ScheduledEventContext {
-  logger: Logger;
-}
-
-export interface ScheduledEventOptions {
-  name: string;
-  buildString?: string;
-}
-
-export type ScheduledEventHandler = (
-  context: ScheduledEventContext
-) => Promise<[number, AnyJson]>;
