+++ jobs/src/util/logger/index.ts
@@ -1,0 +1,4 @@
+export { default as Logger } from './logger';
+export { default as SimpleLogger } from './simpleLogger';
+
+export type { default as ILogger } from './type';
